+++ src/pages/view-test-result/cat-b/components/vehicle-details-card/vehicle-details-card.cat-b.model.ts
