--- src/pages/view-test-result/cat-b/components/vehicle-details-card/vehicle-details-card.scss
@@ -1,3 +1,0 @@
-vehicle-details-card {
-
-}
