+++ src/pages/view-test-result/cat-b/components/view-test-result.cat-b.components.module.ts
@@ -1,14 +1,14 @@
 import { NgModule } from '@angular/core';
 import { CommonModule } from '@angular/common';
 import { IonicModule } from 'ionic-angular';
-import { VehicleDetailsCardComponent } from './vehicle-details-card/vehicle-details-card';
+import { VehicleDetailsCardCatBComponent } from './vehicle-details-card/vehicle-details-card.cat-b';
 import { DebriefCardComponent } from './debrief-card/debrief-card';
 import { ComponentsModule } from '../../../../components/common/common-components.module';
 import { ViewTestResultComponentsModule } from '../../components/view-test-result.components.module';
 
 @NgModule({
   declarations: [
-    VehicleDetailsCardComponent,
+    VehicleDetailsCardCatBComponent,
     DebriefCardComponent,
   ],
   imports: [
@@ -18,7 +18,7 @@ import { ViewTestResultComponentsModule } from '../../components/view-test-resul
     ViewTestResultComponentsModule,
   ],
   exports: [
-    VehicleDetailsCardComponent,
+    VehicleDetailsCardCatBComponent,
     DebriefCardComponent,
   ],
 })
