+++ src/pages/view-test-result/cat-b/view-test-result.cat-b.page.html
@@ -24,7 +24,7 @@
       <rekey-reason-card [data]='testResult.rekeyReason'></rekey-reason-card>
     </div>
     <examiner-details-card [data]='getExaminerDetails()'></examiner-details-card>
-    <vehicle-details-card [data]='getVehicleDetails()'></vehicle-details-card>
+    <vehicle-details-cat-b-card [data]='getVehicleDetails()'></vehicle-details-cat-b-card>
     <debrief-card [data]='getDebrief()'></debrief-card>
     <test-summary-card
       [accompaniment]='testResult.accompaniment'
